% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bins.R
\name{get_bins}
\alias{get_bins}
\title{Get Bins for a Numerical Vector}
\usage{
get_bins(x, bins)
}
\arguments{
\item{x}{a numerical vector}

\item{bins}{a single number (greater than or equal to 2) giving the number of intervals into which x is to be cut.}
}
\value{
a vector of bins
}
\description{
Get Bins for a Numerical Vector
}
