% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bins.R
\name{get_bins_df}
\alias{get_bins_df}
\title{Title}
\usage{
get_bins_df(data, x, bins)
}
\arguments{
\item{data}{a data.frame}

\item{x}{a numerical variable's names that needs to break its values in to bins}

\item{bins}{a single number (greater than or equal to 2) giving the number of intervals into which x is to be cut.}
}
\value{
a vector of bins
}
\description{
Title
}
