% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_correlation.R
\name{get_correlation}
\alias{get_correlation}
\title{Calculate Correlation}
\usage{
get_correlation(data = NULL, ncore = NULL, parallel = TRUE)
}
\arguments{
\item{data}{input data.frame for calculating correlations between variables}

\item{ncore}{number of cores used. Default is to use available cores - 1.}

\item{parallel}{set to TRUE for parallel processing. Number of cores is determined by the ncore parameter.}
}
\value{
a data.frame with variables and their correlations
}
\description{
Calculate Correlation
}
