% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_redundant_vars.R
\name{get_redundant_vars}
\alias{get_redundant_vars}
\title{Find redundant variables based on the correlation data.frame}
\usage{
get_redundant_vars(data, V1, V2, correlation_var, cutoff = 0.7)
}
\arguments{
\item{data}{correlation data.frame from get_correlation function}

\item{V1}{the column name of first variable in the input correlation data.frame}

\item{V2}{the column name of second variable in the input correlation data.frame}

\item{correlation_var}{the column name of variable that measures the correlations between two variables.}

\item{cutoff}{the value of correlation that considered to be highly correlated between two variables.}
}
\value{
a character vectors
}
\description{
Find redundant variables based on the correlation data.frame
}
