% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_responders.R
\name{get_responders}
\alias{get_responders}
\title{calculate number of responders in each bin from the results for WOE table from Information package}
\usage{
get_responders(IV_row, WOE, N_total, N_i)
}
\arguments{
\item{IV_row}{IV of rows from WOE function in Information package}

\item{WOE}{Weight of Evidence value of each row from WOE function in Information package}

\item{N_total}{total records}

\item{N_i}{records in bin i}
}
\value{
a number of responders
}
\description{
calculate number of responders in each bin from the results for WOE table from Information package
}
