% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_response_profile.R
\name{get_response_profile}
\alias{get_response_profile}
\title{Create Response Profile Chart for the Selected Variable (Predictor)}
\usage{
get_response_profile(data, variable_name, y_name, y2_name, title)
}
\arguments{
\item{data}{input data.frame including decile and responder information}

\item{variable_name}{variable name for response profile}

\item{y_name}{primary y axis title}

\item{y2_name}{secondary y axis title}

\item{title}{chart title}
}
\value{
response profile plot
}
\description{
Create Response Profile Chart for the Selected Variable (Predictor)
}
