% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTile.R
\name{getTile}
\alias{getTile}
\title{Get Tiles for the scoring Process}
\usage{
getTile(ScoreFile, ScoredVariable, BestTile1 = FALSE, TileNum = 10)
}
\arguments{
\item{ScoreFile}{input data.frame with scores}

\item{ScoredVariable}{the names of score variable}

\item{BestTile1}{specify which tile corresponding to high scores}

\item{TileNum}{numbers of tiles}
}
\value{
a data.frame with bins
}
\description{
Get Tiles for the scoring Process
}
