% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ImportantVariables}
\alias{ImportantVariables}
\title{Variable importance dataset from a model}
\format{
A data frame with 36 rows and 3 variables:
\itemize{
\item Variable: the names of predictors in a model
\item Importance: the percentage of importance
\item Sources: the sources of the predictors
}
}
\usage{
ImportantVariables
}
\description{
The vaiable importance data is from a h2o xgboost model.
It contains 3 variables and 36 records.
}
\keyword{datasets}
