% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IRRCRR_plot.R
\name{IRRCRR_plot}
\alias{IRRCRR_plot}
\title{Title}
\usage{
IRRCRR_plot(preds.MarketingCharts_model, adjust = 1, Interval = 0.01)
}
\arguments{
\item{preds.MarketingCharts_model}{input data.frame with  IRR and CRR}

\item{adjust}{a number use to adjust line position}

\item{Interval}{a number for interval between breaks in y axis}
}
\value{
IRR and CRR plot
}
\description{
Title
}
