% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{IV_WOE_Table}
\alias{IV_WOE_Table}
\title{Predictors response profiles dataset from WOE analysis}
\format{
A data frame with 335 rows and 7 variables:
\itemize{
\item Sources: the sources of the predictors
\item Variable: the names of predictors in a model
\item Variable_Bin: the bin of numerical variables or the value of catogorical variables
\item Population: the total records in the bin or values
\item IV: the information value from WOE analysis
\item Responders: the number of responders in that bin or values
\item Conversion_Rate: the conversion rate in that bin or values
}
}
\usage{
IV_WOE_Table
}
\description{
The data is from Information package WOE analysis
It contains 7 variables and 335 records.
}
\keyword{datasets}
