% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_plot.R
\name{lift_plot}
\alias{lift_plot}
\title{Lift over Random Plot}
\usage{
lift_plot(
  preds.MarketingCharts.df,
  Interval = 0.25,
  color = "#0066CC",
  type = "single",
  color_legend = NULL,
  color_V = NULL
)
}
\arguments{
\item{preds.MarketingCharts.df}{nput data.frame with  Lift}

\item{Interval}{a number for interval between breaks in y axis}

\item{color}{color for line}

\item{type}{a string if it is "single" then create lift plot for one model otherwise create lift plot for multiple models}

\item{color_legend}{legend for color}

\item{color_V}{vector of colors}
}
\value{
lift plot
}
\description{
Lift over Random Plot
}
