% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_plot.R
\name{lift_plot_x}
\alias{lift_plot_x}
\title{Group Level Lift over Random Plot}
\usage{
lift_plot_x(
  Data_LiftPlot,
  Group_Variable,
  Interval = 0.25,
  color = "#0066CC",
  color_legend = NULL,
  color_V = NULL
)
}
\arguments{
\item{Data_LiftPlot}{input data.frame with  lift and group variable}

\item{Group_Variable}{variable name for  grouping}

\item{Interval}{a number for interval between breaks in y axis}

\item{color}{color for line}

\item{color_legend}{legend for color}

\item{color_V}{vector of colors}
}
\value{
lift plot
}
\description{
Group Level Lift over Random Plot
}
