% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketingCharts.R
\name{MarketingCharts}
\alias{MarketingCharts}
\title{Marketing Gains Charts}
\usage{
MarketingCharts(pred, test, type, NBin = 10)
}
\arguments{
\item{pred}{a vector of the probability of the positive class (response).}

\item{test}{a vector of the real observation.}

\item{type}{a string of CumulativeGain","Gain","Lift"}

\item{NBin}{a number of bins}
}
\value{
a data.frame of gains charts
}
\description{
Marketing Gains Charts
}
