% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testing_frame}
\alias{testing_frame}
\title{dataset with response variable y and predictors and models scores}
\format{
A data frame with 11,701 rows and 43 variables:
\itemize{
\item y: equals 1 if the person response the marketing offer, and 0 if the person didn't response.
\item Age: age of the person.
\item glm: score from glm model.
\item rf: score from rf model.
\item gbm: score from gbm model.
\item automl: score from h2o automl model.
\item xgboost: score from xgboost model.
\item Population Density: demographic variable
\item Health and Beauty: demographic variable
\item Home Length of Residence - Actual (RP)1: demographic variable
\item Other variables are from demographic data
}
}
\usage{
testing_frame
}
\description{
The data combined with modeling data and sores from a predictive models(glm, rf, gbm, automl and xgboost).
It contains one response variable y and 37 predictive variables and 11,701 records. In addition, it contains scores
from five models (glm, rf, gbm, automl and xgboost)
}
\keyword{datasets}
