% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_importance_plot.R
\name{variable_importance_plot}
\alias{variable_importance_plot}
\title{Create a variable importance plot}
\usage{
variable_importance_plot(
  Data_Plot,
  variable_name,
  importance_name,
  bar_fill = "#4f81bd",
  subtitle = NULL,
  caption = NULL,
  source_name = NULL,
  source_colors = NULL,
  num_of_features = 10
)
}
\arguments{
\item{Data_Plot}{input data.frame for plot (it includes variable and it’s importance ).}

\item{variable_name}{the column name in the input data.frame that includes the variables of interest.}

\item{importance_name}{The column name in the input data.frame that includes measures of variable importance.}

\item{bar_fill}{the color for bar chart.}

\item{subtitle}{The subtitle of the plot.}

\item{caption}{The caption of the plot.}

\item{source_name}{the column name in the input data.frame that includes source of the variables of analysis.}

\item{source_colors}{the vector of colors for the different source. (The source_name must specified)}

\item{num_of_features}{the number of variables display in the plot.}
}
\value{
variable importance plot
}
\description{
Create a variable importance plot
}
